power: bring up qc, wire it into tests/c/mini and the top-level build
Fixes two compile errors in compilers/qc/mkfile, both the exact same
class already hit and fixed in zc's own bring-up (523eadaad):
align()/maxround() needed int32 not long to match compilers/cck/cc.h's
prototypes, and gc.h's swcmp/hintabsize needed the same const-qualified
prototype / plain-extern fixes zc's gc.h already has.

Adds mkfiles/power/mkfile (CC=qc, AS=qa, LD="ql -0 -H7", AR=iar) and
compilers/qc to the top-level mkfile's DIRS (assemblers/qa and
linkers/ql were already there), so `mk objtype=power` and the
top-level build both know about qc.

Adds tests/c/mini/{start,xwrite,xexit}_power.s, derived from reading
qc's actual codegen with `qc -S` on a throwaway two-arg function
(first arg in R3, further args at argname+4(FP)) rather than guessing.

Getting `ql -L` to link the resulting libmini.a surfaced a real gap:
lib_toolchain/libmach had no dispatch entry at all for power's '.q'
objects (q.c/qobj.c/qdb.c were imported alongside qa/ql/qc but never
wired up or even added to libmach's own OFILES), so iar could not
build a symbol table for the archive -- same "first entry not symbol
header" error zl hit for the identical reason during its own bring-up.
Added the missing MPOWER/FPOWER/APOWER/APOWER64/ObjPower enum values
(include/arch/mach.h), the obj.c dispatch row, the setmach.c machines[]
entry, the three missing OFILES, and fixed two leftover
absolute/relative include paths in q.c and qobj.c.

Verified: helloc.exe (real qc-compiled C) now runs under qemu-ppc with
correct output and exit(0), byte-matching every other arch in
tests/c/mini's ARCHS. Rebuilt and re-verified debuggers/acid,
machines/5i, machines/vi (all link libmach.a) still pass their own
smoke tests after the shared mach.h/libmach changes.

Also adds etc/acid/power, a real per-arch acid startup script that was
sitting untracked in the working tree -- evidently part of the
original Plan 9 4th edition import but missed by cf54e8364's `git add`.

Records the full bring-up in docs/claude_notes/notes_arch_power.txt.

+++ compilers/qc/swt.c
@@ -548,10 +548,10 @@ doubled(Type *t)
 	return 0;
 }
 
-long
-align(long i, Type *t, int op)
+int32
+align(int32 i, Type *t, int op)
 {
-	long o;
+	int32 o;
 	Type *v;
 	int w, pc;
 
@@ -632,8 +632,8 @@ align(long i, Type *t, int op)
 	return o;
 }
 
-long
-maxround(long max, long v)
+int32
+maxround(int32 max, int32 v)
 {
 	int w;
